+++ src/jmnavpilot/SimpleAutopilot.java
@@ -18,24 +18,24 @@ public class SimpleAutopilot implements Pilot {
     }
 
     @Override
-    public float[] getControls() {
+    public double[] getControls() {
 
         // P control Roll stabilizer. K = 0.5
-//        float roll = -0.3f * cs.getRoll() - 0.75f * cs.getSensors().getzRate();
-//        float roll = -0.25f * cs.getRoll() - 0.125f * cs.getYaw();
-        float roll = -0.2f * cs.getRoll();
-//        float roll = 0.2f * cs.getSensors().getyAccel(); // Roll ??
+//        double roll = -0.3f * cs.getRoll() - 0.75f * cs.getSensors().getzRate();
+//        double roll = -0.25f * cs.getRoll() - 0.125f * cs.getYaw();
+        double roll = -0.2f * cs.getRoll();
+//        double roll = 0.2f * cs.getSensors().getyAccel(); // Roll ??
         roll = Math.min(1, Math.max(-1, roll));
 
         // Hard takeoff, level flight alt/thrust
-        float pitch = -0.0f;
-        float thrust = -0.07f * (cs.getAlt() - 25);
+        double pitch = -0.0f;
+        double thrust = -0.07f * (cs.getAlt() - 25);
         thrust = Math.min(0.77f, Math.max(0, 0.5f + thrust / 2));
 //        if (cs.getAlt() < 1) {
 //            pitch = -1f;
 //            thrust = 1f;
 //        }
-        float[] result = {roll, pitch, thrust};
+        double[] result = {roll, pitch, thrust};
         return result;
     }
 }
